fixes for msvc level4 warnings
git-svn-id: http://svn.boost.org/svn/boost/trunk@20198 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/boost/date_time/time_clock.hpp
@@ -2,7 +2,7 @@
 #define DATE_TIME_TIME_CLOCK_HPP___
 /* Copyright (c) 2002 CrystalClear Software, Inc.
  * Disclaimer & Full Copyright at end of file
- * Author: Jeff Garland 
+ * Author: Jeff Garland, Bart Garst 
  */
 
 /*! @file time_clock.hpp
@@ -47,8 +47,8 @@ namespace date_time {
     static time_type create_time(::std::tm* current)
     {
       date_type d(current->tm_year + 1900, 
-                  current->tm_mon + 1, 
-                  current->tm_mday);
+                  static_cast<unsigned short>(current->tm_mon + 1), 
+                  static_cast<unsigned short>(current->tm_mday));
       time_duration_type td(current->tm_hour,
                             current->tm_min,
                             current->tm_sec);
